#5 revert BaseMethodIdentifier
